Fix for class-loader libz and pcre link issue
The class-loader package fails if the pre-built binary of Poco
library is used. The PocoFoundationTargets.cmake file coming
with the pre-built binary have hardcoded paths to libz and libpcre
which makes the cross-compilation to fail.
This fix removes Poco library from the apt-get list and forces
to build it inside the poco_vendor ros2 packages.

Change-Id: Iae4f04532bf452735c35acd4d287c2631e7ad8a1
Signed-off-by: Louis Mayencourt <louis.mayencourt@arm.com>