Refactor workflow to only report metrics once
Before this commit, one metric was emitted per entry in the workflow
matrix. This is a problem because as it makes our dashboards harder to
read (different repositories have different expected values).

This refactors the workflow to only emit the metric once. If the
workflow fails on either Mac or Windows, 0 is emitted. If it succeeds on
both platforms, 1 is emitted.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>