Add merge-install and explicit colcon defaults to the Tutorial
A recent change broke the default-`defaults.yaml` usage, so it needs to be explicitly specified until that is fixed. Additionally, the "deploying to a target system" instructions require a merged install for `rosdep install --from-paths install/share` to work, because otherwise the `share` directories are distributed across each package's install-subdirectory.